# backend/openapi.yaml
openapi: 3.0.3
info:
  title: 故事森林 API
  description: |
    ## 故事森林后端 API 文档
    
    面向 3-10 岁儿童的互动式故事阅读平台后端 API。
    
    ### 认证说明
    大部分接口需要 JWT Token 认证，在请求头中携带：
    ```
    Authorization: Bearer <your_access_token>
    ```
    
    ### 响应格式
    所有接口统一返回格式：
    ```json
    {
      "code": 0,
      "message": "success",
      "data": {}
    }
    ```
    
    ### 错误码说明
    - `0`: 成功
    - `400`: 请求参数错误
    - `401`: 未授权/Token 无效
    - `403`: 权限不足
    - `404`: 资源不存在
    - `500`: 服务器内部错误
  version: 1.0.0
  contact:
    name: 故事森林团队
    email: support@storyforest.com
  license:
    name: Copyright © 2026 故事森林
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
  - url: https://api.storyforest.com/api
    description: 生产环境
  - url: https://test-api.storyforest.com/api
    description: 测试环境
...
paths:
...
  /stories:
    get:
      tags:
        - 故事管理
      summary: 获取故事列表
      description: 分页获取故事列表，支持排序
      operationId: getStories
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: 页码（从 0 开始）
        - name: size
          in: query
          schema:
            type: integer
            default: 10
          description: 每页数量
        - name: sort
          in: query
          schema:
            type: string
            default: playCount
          description: 排序字段
        - name: direction
          in: query
          schema:
            type: string
            default: desc
            enum: [asc, desc]
          description: 排序方向
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Page_Story'
...
  /stories/{id}:
    get:
      tags:
        - 故事管理
      summary: 获取故事详情
      description: 根据故事 ID 获取详细信息
      operationId: getStoryDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Story'
        '404':
          description: 故事不存在
  /stories/{id}/play:
    post:
      tags:
        - 故事管理
      summary: 增加播放次数
      description: 记录故事播放，增加播放计数
      operationId: incrementPlayCount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 成功
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证，格式：Bearer <token>
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 响应码 (0=成功，其他=失败)
          example: 0
        message:
          type: string
          description: 响应消息
          example: success
        data:
          type: object
          description: 响应数据
          nullable: true
...
    ApiResponse_Story:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Story'
    ApiResponse_Page_Story:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Page_Story'
...
    Story:
      type: object
      properties:
        id:
          type: integer
          description: 故事 ID
          example: 1
        title:
          type: string
          description: 故事标题
          example: "小红帽"
        subtitle:
          type: string
          description: 副标题
          example: "经典童话故事"
        coverImage:
          type: string
          description: 封面图 URL
        description:
          type: string
          description: 故事简介
        ageGroup:
          type: string
          description: 年龄段
          example: "5-6 岁"
        category:
          type: string
          description: 主题分类
          example: "童话故事"
        durationType:
          type: string
          description: 时长分类
          example: "short"
          enum: [short, medium, long]
        readingMinutes:
          type: integer
          description: 预计阅读时长（分钟）
          example: 5
        chapterCount:
          type: integer
          description: 章节数量
          example: 3
        playCount:
          type: integer
          description: 播放次数
          example: 1000
        favoriteCount:
          type: integer
          description: 收藏次数
          example: 200
        averageRating:
          type: number
          description: 平均评分
          example: 4.5
        isRecommended:
          type: boolean
          description: 是否推荐
          example: true
        isVip:
          type: boolean
          description: 是否 VIP 专享
          example: false
        status:
          type: string
          description: 状态
          example: PUBLISHED
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
    Page_Story:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Story'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          description: 总元素数
        totalPages:
          type: integer
          description: 总页数
        size:
          type: integer
          description: 每页大小
        number:
          type: integer
          description: 当前页码
        first:
          type: boolean
          description: 是否第一页
        last:
          type: boolean
          description: 是否最后一页
        empty:
          type: boolean
          description: 是否为空
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
          description: 页码
        pageSize:
          type: integer
          description: 每页大小
        sort:
          type: array
          items:
            $ref: '#/components/schemas/Sort'
    Sort:
      type: object
      properties:
        property:
          type: string
          description: 排序属性
        direction:
          type: string
          description: 排序方向
          enum: [ASC, DESC]